requirements.txt
# ============================================================================
# CORE DEPENDENCIES (Production)
# ============================================================================

# Deep Learning & PyTorch
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
ultralytics==8.2.0

# Computer Vision & ANPR
opencv-python==4.9.0.80
pillow==10.1.0
scikit-image==0.22.0
easyocr==1.7.0
paddleocr==2.7.0.3
pytesseract==0.3.10

# Data Handling & Analytics
numpy==1.26.0
pandas==2.2.0
scipy==1.12.0

# Machine Learning & Tracking
scikit-learn==1.4.0
deep-sort-realtime==1.3.2
filterpy==1.4.5
mlflow==2.11.1

# Reinforcement Learning & Control
stable-baselines3==2.2.1
gymnasium==0.29.1

# Web Framework & API
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.0
pydantic-settings==2.1.0
websockets==12.0
aiofiles==23.2.1
python-multipart==0.0.6

# Database & Migrations
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.12.0
redis==5.0.1

# Notifications & Alerts
twilio==8.9.0
sendgrid==6.10.0

# Infrastructure & Monitoring
prometheus-client==0.19.0
psutil==6.0.0
PyYAML==6.0.1
python-dotenv==1.0.0

# Performance Optimization
numba==0.58.1

# Utilities
loguru==0.7.2
sentry-sdk==1.39.1
requests==2.31.0
tqdm==4.67.0
cryptography==41.0.7

# Visualization & Dashboard
matplotlib==3.8.3
plotly==5.19.0
streamlit==1.37.0

# E-Challan PDF Generation (Feature 5)
reportlab==4.1.0
qrcode[pil]==7.4.2
